requirements.txt
python-barcode
pyzbar
qrcode
Pillow